Move DataModel to Core.
