+++ BugReport/Program.cs
@@ -3,10 +3,10 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
 using BugReport.CommandLine;
-using BugReport.DataModel;
 using BugReport.Reports;
 using BugReport.Reports.EmailReports;
 using BugReport.Util;
+using GitHubBugReport.Core.DataModel;
 using GitHubBugReport.Core.Issues.Models;
 using GitHubBugReport.Core.Issues.Services;
 using GitHubBugReport.Core.Storage.Services;
