+++ BugReport/Reports/EmailReport/AlertReport_NeedsMSResponse.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Web;
-using BugReport.DataModel;
 using BugReport.Util;
 using GitHubBugReport.Core.Issues.Models;
